requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
httpx==0.25.1
Pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiomysql==0.2.0
prometheus-fastapi-instrumentator==6.1.0

# 新增补齐依赖
python-jose==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
redis[async]==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1